+++ packages/core/src/observable/types.ts
@@ -1,7 +1,11 @@
+/** Type of {@link Observer.next} */
 export type ObserverNext<T> = (value: T) => void
+/** Type of {@link Observer.error} */
 export type ObserverError = (e: any) => void
+/** Type of {@link Observer.complete} */
 export type ObserverComplete = () => void
 
+/** Observer mimics Observer from ECMAScript TC39 Observable proposal */
 export interface Observer<T> {
   next: ObserverNext<T>
   error: ObserverError
@@ -10,10 +14,10 @@ export interface Observer<T> {
 
 /**
  * An observable that aligns with the
- * [TC39 observable proposal](https://github.com/tc39/proposal-observable) and
+ * {@link https://github.com/tc39/proposal-observable | TC39 observable proposal} and
  * can be consumed by other observable libraries like
- * [rxjs](https://github.com/ReactiveX/rxjs) or
- * [zen-observable](https://github.com/zenparsing/zen-observable).
+ * {@link https://github.com/ReactiveX/rxjs | rx js} or
+ * {@link https://github.com/zenparsing/zen-observable | zen-observable}.
  */
 export interface Observable<T> {
   subscribe(): Subscription
@@ -28,6 +32,7 @@ export interface Observable<T> {
   /* [Symbol.observable](): Observable<T> */
 }
 
+/** Subscription mimics Subscription from ECMAScript TC39 Observable proposal */
 export interface Subscription {
   readonly closed: boolean
   unsubscribe(): void