+++ packages/core/src/query/FluxTableMetaData.ts
@@ -4,7 +4,7 @@ import {IllegalArgumentError} from '../errors'
 const identity = (x: string): any => x
 /**
  * A dictionary of serializers of particular types returned by a flux query.
- * See https://v2.docs.influxdata.com/v2.0/reference/syntax/annotated-csv/#valid-data-types
+ * See {@link https://v2.docs.influxdata.com/v2.0/reference/syntax/annotated-csv/#valid-data-types }
  */
 export const typeSerializers: Record<ColumnType, (val: string) => any> = {
   boolean: (x: string): any => x === 'true',
@@ -17,7 +17,7 @@ export const typeSerializers: Record<ColumnType, (val: string) => any> = {
   duration: identity,
 }
 /**
- * Represents metadata of a [flux table](http://bit.ly/flux-spec#table).
+ * Represents metadata of a {@link http://bit.ly/flux-spec#table | flux table}.
  */
 export default class FluxTableMetaData {
   /**