+++ packages/core/src/query/index.ts
@@ -3,5 +3,9 @@ export {
   typeSerializers,
 } from './FluxTableMetaData'
 export {default as FluxResultObserver} from './FluxResultObserver'
-export {default as FluxTableColumn, ColumnType} from './FluxTableColumn'
+export {
+  default as FluxTableColumn,
+  ColumnType,
+  FluxTableColumnLike,
+} from './FluxTableColumn'
 export * from './flux'