+++ packages/core/src/util/currentTime.ts
@@ -1,5 +1,3 @@
-import {WritePrecision} from '../options'
-
 declare let process: any
 const zeroPadding = '000000000'
 let useHrTime = false
@@ -77,20 +75,22 @@ function seconds(): string {
  * depending on the js platform in use.
  */
 export const currentTime = Object.freeze({
-  [String(WritePrecision.s)]: seconds,
-  [String(WritePrecision.ms)]: millis,
-  [String(WritePrecision.us)]: micros,
-  [String(WritePrecision.ns)]: nanos,
-  seconds,
-  millis,
-  micros,
-  nanos,
+  s: seconds as () => string,
+  ms: millis as () => string,
+  us: micros as () => string,
+  ns: nanos as () => string,
+  seconds: seconds as () => string,
+  millis: millis as () => string,
+  micros: micros as () => string,
+  nanos: nanos as () => string,
 })
 
+/**
+ * dateToProtocolTimestamp provides converters for JavaScript Date to InfluxDB Write Protocol Timestamp. Keys are supported precisions.
+ */
 export const dateToProtocolTimestamp = {
-  [String(WritePrecision.s)]: (d: Date): string =>
-    `${Math.floor(d.getTime() / 1000)}`,
-  [String(WritePrecision.ms)]: (d: Date): string => `${d.getTime()}`,
-  [String(WritePrecision.us)]: (d: Date): string => `${d.getTime()}000`,
-  [String(WritePrecision.ns)]: (d: Date): string => `${d.getTime()}000000`,
+  s: (d: Date): string => `${Math.floor(d.getTime() / 1000)}`,
+  ms: (d: Date): string => `${d.getTime()}`,
+  us: (d: Date): string => `${d.getTime()}000`,
+  ns: (d: Date): string => `${d.getTime()}000000`,
 }