+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 1. [#218](https://github.com/influxdata/influxdb-client-js/pull/218): Add default documentation to body property in generated APIs.
 1. [#220](https://github.com/influxdata/influxdb-client-js/pull/220): Generate API constructor with InfluxDB parameter.
 1. [#222](https://github.com/influxdata/influxdb-client-js/pull/222): Improve and validate code documentation.
+1. [#224](https://github.com/influxdata/influxdb-client-js/pull/224): Generate API documentation for gh-pages.
 
 ## 1.5.0 [2020-07-17]